fix: states moved to parent component
